Nits

+++ prover/src/opcodes/binary_ops/b32.rs
@@ -164,7 +164,7 @@ impl TableFiller<ProverPackedField> for AndiTable {
                 dst_abs[i] = event.fp.addr(event.dst as u32);
                 dst_val_unpacked[i] = event.dst_val as u16;
                 src_abs[i] = event.fp.addr(event.src as u32);
-                src_val_unpacked[i] = event.src_val;
+                src_val_unpacked[i] = event.src_val as u16;
             }
         }
         let cpu_rows = rows.map(|event| CpuGadget {
